Add support for healthchecks (#2)
Amazon's elastic load balancer (ELB) checks /health to see if our
service is healthy, which was failing before this change.

This adds a new middleware to support that additional endpoint.  The
reason I didn't use Servant to do this is that all of the
Servant-generated endpoints go through the Slack signature verification
(and the health check is not signed by Slack, obviously).