Fix Ada outside of DMS (#7)
The change in #6 broke Ada outside of DMs because she attempts to fetch
thread history for the incoming message, but that fetch only works for
scopes that she supports (currently just DMs) and fails in other
scopes (e.g. channels).  This fixes that to gracefully falling back
to an empty thread context when she's missing the required scopes.