Embed individual documents if a batch embedding fails
I learned the hard way that the change in #21 is not exactly enough to
fix the resilience of indexing because if one document in a batch fails
to embed then the whole batch fails.

This change causes the whole batch to be split up into individual
documents when that happens, which not only limits the blast radius of
an embedding failure but also improves the error message, too (because
now we know exactly which document failed).